pyproject.toml
[project]
name = "stock-ui"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "flask>=3.0",
  "gunicorn>=21.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "requests>=2.32",
  "python-dateutil>=2.9",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
